Moved decorator to another lib

+++ Dotnet.Homeworks.Features/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,24 +1,30 @@
 ﻿using Dotnet.Homeworks.Domain.Abstractions.Repositories;
-using Dotnet.Homeworks.Infrastructure.Cqrs.Commands;
+using Dotnet.Homeworks.Features.Decorators;
 using Dotnet.Homeworks.Infrastructure.UnitOfWork;
+using Dotnet.Homeworks.Infrastructure.Validation.PermissionChecker;
 using Dotnet.Homeworks.Shared.Dto;
+using FluentValidation;
 
 namespace Dotnet.Homeworks.Features.Users.Commands.DeleteUser;
 
-public class DeleteUserCommandHandler : ICommandHandler<DeleteUserCommand>
+public class DeleteUserCommandHandler : CqrsDecorator<DeleteUserCommand, Result>
 {
     private readonly IUserRepository _userRepository;
     private readonly IUnitOfWork _unitOfWork;
 
-
-    public DeleteUserCommandHandler(IUserRepository userRepository, IUnitOfWork unitOfWork)
+    public DeleteUserCommandHandler(IUserRepository userRepository, IUnitOfWork unitOfWork, IEnumerable<IValidator<DeleteUserCommand>> validators, IPermissionCheck<DeleteUserCommand>? permissionCheck) : base(validators, permissionCheck)
     {
         _userRepository = userRepository;
         _unitOfWork = unitOfWork;
     }
 
-    public async Task<Result> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+    public override async Task<Result> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
+        var pipelineResult = await base.Handle(request, cancellationToken);
+
+        if (pipelineResult.IsFailure)
+            return pipelineResult;
+
         try
         {
             await _userRepository.DeleteUserByGuidAsync(request.Guid, cancellationToken);
